# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Pam English
# during dev, should point to your local machine
host: localhost:3005
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
  - application/octet-stream
# format of the responses to the client (Accepts)
produces:
  - "application/json"
  - "application/xml"
  - "application/x-www-form-urlencoded"
paths:
...
  /users:
    x-swagger-router-controller: users
    post:
      summary: "Add a new user(dev.need validation check)"
      description: Create user
      operationId: createUser
      parameters: [
        {
          name: name,
          in: formData,
          description: '이름',
          required: true,
          type: string
        }, {
          name: password,
          in: formData,
          description: '비밀번호',
          required: false,
          type: string
        }, {
          name: email,
          in: formData,
          description: '이메일',
          required: true,
          type: string
        }, {
          name: mobile,
          in: formData,
          description: '핸드폰번호(숫자만입력)',
          required: false,
          type: string
        }, {
          name: nick,
          in: formData,
          description: '별명',
          required: false,
          type: string
        }, {
          name: levelId,
          in: formData,
          description: 'LEVEL의 ID값',
          required: true,
          type: string
        }, {
          name: snsProvider,
          in: formData,
          description: 'sns provider(kakaotalk, naver, facebook, etc.)',
          required: false,
          type: string
        }, {
          name: snsId,
          in: formData,
          description: 'sns id',
          required: false,
          type: string
        }
      ]
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsersResponse"
        "409":
          description: Conflict
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UsersResponse:
    type: object
    properties:
      success:
        type: boolean
      userId:
        type: integer
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string